test: add failing test case for projection that removes sorting

+++ ibis/backends/flink/tests/test_ddl.py
@@ -32,11 +32,8 @@ def awards_players_csv_connector_configs():
 
 
 def test_list_tables(con):
-    assert len(con.list_tables()) == 4
-    assert (
-        len(con.list_tables(catalog="default_catalog", database="default_database"))
-        == 4
-    )
+    assert len(con.list_tables())
+    assert con.list_tables(catalog="default_catalog", database="default_database")
 
 
 def test_create_table_from_schema(
@@ -47,38 +44,26 @@ def test_create_table_from_schema(
         schema=awards_players_schema,
         tbl_properties=awards_players_csv_connector_configs,
     )
-    assert len(con.list_tables()) == 5
     assert temp_table in con.list_tables()
     assert new_table.schema() == awards_players_schema
 
 
-def test_drop_table(
-    con, awards_players_schema, temp_table, awards_players_csv_connector_configs
+@pytest.mark.parametrize("temp", [True, False])
+def test_create_table(
+    con, awards_players_schema, temp_table, awards_players_csv_connector_configs, temp
 ):
     con.create_table(
         temp_table,
         schema=awards_players_schema,
         tbl_properties=awards_players_csv_connector_configs,
+        temp=temp,
     )
-    assert len(con.list_tables()) == 5
-    con.drop_table(temp_table)
-    assert len(con.list_tables()) == 4
-    assert temp_table not in con.list_tables()
-
-
-def test_temp_table(
-    con, awards_players_schema, temp_table, awards_players_csv_connector_configs
-):
-    con.create_table(
-        temp_table,
-        schema=awards_players_schema,
-        tbl_properties=awards_players_csv_connector_configs,
-        temp=True,
-    )
-    assert len(con.list_tables()) == 5
     assert temp_table in con.list_tables()
-    with pytest.raises(Py4JJavaError):
-        con.drop_table(temp_table)
-    con.drop_table(temp_table, temp=True)
-    assert len(con.list_tables()) == 4
+
+    if temp:
+        with pytest.raises(Py4JJavaError):
+            con.drop_table(temp_table)
+
+    con.drop_table(temp_table, temp=temp)
+
     assert temp_table not in con.list_tables()
